feat : 유튜브 검색결과 저장 구현 중 임시 커밋 - 재생목록, 채널정보 저장 완료 - 영상목록, 댓글목록 구현 중

+++ src/main/java/com/steelrain/lilac/batch/config/APIConfig.java
@@ -12,6 +12,6 @@ public class APIConfig {
 
     public APIConfig(String youtubeKey, String threshold){
         this.youtubeKey = youtubeKey;
-        this.threshold = Float.valueOf(threshold);
+        this.threshold = Float.valueOf(threshold).floatValue();
     }
 }
